+++ src/ECS/SpriteComponent.h
@@ -60,8 +60,8 @@ public:
 
 		destRect.x = static_cast<int>(transform->position.x) - Game::camera.x;
 		destRect.y = static_cast<int>(transform->position.y) - Game::camera.y;
-		destRect.w = transform->width * transform->scale;
-		destRect.h = transform->height * transform->scale;
+		destRect.w = static_cast<int>(transform->width * transform->scale);
+		destRect.h = static_cast<int>(transform->height * transform->scale);
 	}
 
 	void draw() override
